test: 테스트코드 오류 수정
