+++ src/main/java/server/poptato/todo/api/TodoController.java
@@ -11,6 +11,8 @@ import server.poptato.todo.application.response.BacklogListResponseDto;
 import server.poptato.todo.application.response.TodayListResponseDto;
 import server.poptato.user.resolver.UserId;
 
+import java.time.LocalDate;
+
 @RestController
 @RequiredArgsConstructor
 public class TodoController {
@@ -20,7 +22,8 @@ public class TodoController {
             @UserId Long userId,
             @RequestParam(value = "page", defaultValue = "0") int page,
             @RequestParam(value = "size", defaultValue = "8") int size){
-        TodayListResponseDto todayListResponse = todoService.getTodayList(userId, page, size);
+        LocalDate todayDate = LocalDate.now();
+        TodayListResponseDto todayListResponse = todoService.getTodayList(userId, page, size, todayDate);
         return new BaseResponse<>(todayListResponse);
     }
